% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\docType{package}
\name{gwas.explorer}
\alias{gwas.explorer}
\title{gwas.explorer: A package for qyerying GWAS results from studies supported by
the National Cancer Institute}
\description{
The GWAS Explorer SDK enables users to query GWAS results and
metadata from studies that were supported by the National Cancer Institute.
The SDK currently provides GWAS results for the Prostate, Lung, Colorectal,
and Ovarian Cancer Screening Trial (PLCO).
}
