% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.download_dataset}
\alias{plco.download_dataset}
\title{Downloads the original association results in tsv.gz format to the specified path}
\usage{
plco.download_dataset(phenotype_id, filepath, ...)
}
\arguments{
\item{phenotype_id}{A numeric phenotype id}

\item{sex}{A character vector specifying the path to save the file to}
}
\description{
Downloads the original association results in tsv.gz format to the specified path
}
\examples{
plco.download_dataset(1010, "download.tsv.gz")
}
