% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.get_participants}
\alias{plco.get_participants}
\title{Retrieves aggregate counts for participants. Aggregate counts under 10 are returned as "< 10".}
\usage{
plco.get_participants(phenotype_id, columns = c("value"), precision = 0, ...)
}
\arguments{
\item{columns}{A character vector specifying properties for which to retrieve counts for. Valid properties are: value, ancestry, genetic_ancestry, sex, and age}

\item{precision}{Optional. For continuous phenotypes, a numeric value specifying the -log10(precision) to which values should be rounded to.}
}
\description{
Retrieves aggregate counts for participants. Aggregate counts under 10 are returned as "< 10".
}
\examples{
plco.get_participants(1010)
plco.get_participants(1010, c("value", "age"))
plco.get_participants(2250, c("value"), -1)
}
