% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.get_pca}
\alias{plco.get_pca}
\title{Retrieve PCA coordinates for the specified phenotype and platform}
\usage{
plco.get_pca(phenotype_id, platform, pc_x, pc_y, ...)
}
\arguments{
\item{phenotype_id}{A numeric phenotype id}

\item{platform}{A character vector specifying the platform to retrieve data for}

\item{pc_x}{A numeric value (1-20) specifying the x axis's principal component}

\item{pc_y}{A numeric value (1-20) specifying the y axis's principal component}
}
\value{
A dataframe containing pca coordinates
}
\description{
Retrieve PCA coordinates for the specified phenotype and platform
}
