% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.get_phenotypes}
\alias{plco.get_phenotypes}
\title{Retrieve phenotypes}
\usage{
plco.get_phenotypes(query = NULL, ...)
}
\arguments{
\item{query}{A character vector specifying a term to search for}
}
\value{
If query is specified, returns a list of phenotypes which contain the query term. Otherwise, returns all phenotypes. Note that phenotypes without an import_count do not contain associated gwas summary results
}
\description{
Retrieve phenotypes
}
\examples{
plco.get_phenotypes("cancer")
}
