% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.get_points}
\alias{plco.get_points}
\title{Retrieves sampled variants suitable for visualizing a QQ plot for the specified phenotype, sex, and ancestry}
\usage{
plco.get_points(phenotype_id, sex, ancestry, ...)
}
\arguments{
\item{phenotype_id}{A numeric phenotype id}

\item{sex}{A character vector specifying a sex to retrieve data for}

\item{ancestry}{A character vector specifying ancestries to retrieve data for}
}
\value{
A dataframe containing variants
}
\description{
Retrieves sampled variants suitable for visualizing a QQ plot for the specified phenotype, sex, and ancestry
}
\examples{
plco.get_points(1010, "female", "european")
}
