% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.get_variants}
\alias{plco.get_variants}
\title{Retrieves variants for a specified phenotype, sex, and ancestry}
\usage{
plco.get_variants(
  phenotype_id,
  sex,
  ancestry,
  chromosome = NULL,
  columns = NULL,
  snp = NULL,
  position = NULL,
  position_min = NULL,
  position_max = NULL,
  p_value_min = NULL,
  p_value_max = NULL,
  order_by = "p_value",
  order = "asc",
  offset = 0,
  limit = 1000,
  ...
)
}
\arguments{
\item{phenotype_id}{A numeric phenotype id}

\item{sex}{A character vector specifying a sex to retrieve data for}

\item{ancestry}{A character vector specifying ancestries to retrieve data for}

\item{chromosome}{Optional. A character vector specifying a chromosome to retrieve data for}

\item{columns}{Optional. A character vector specifying properties to include for each variant}

\item{snp}{Optional. A character vector specifying snps to retrieve}

\item{position}{Optional. A numeric value specifying the chromosome position to retrieve data for}

\item{position_min}{Optional. A numeric value specifying the minimum chromosome position for variants}

\item{position_max}{Optional. A numeric value specifying the maximum chromosome position  for variants}

\item{p_value_min}{Optional. A numeric value specifying the minimum p-value for variants}

\item{p_value_max}{Optional. A numeric value specifying the maximum p-value for variants}

\item{order_by}{Optional. A character vector specifying the property to sort variants by}

\item{order}{Optional. A character vector specifying the order to sort variants by. Either "asc" or "desc"}

\item{offset}{Optional. A numeric value specifying the offset from which to begin retrieving variants}

\item{limit}{Optional. A numeric value specifying the maximum number of variants to retrieve}
}
\value{
A dataframe containing variants
}
\description{
Retrieves variants for a specified phenotype, sex, and ancestry
}
\examples{
plco.get_variants(1010, "female", "european", limit=100)
}
