% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plco.R
\name{plco.request}
\alias{plco.request}
\title{Performs a request against the PLCO API}
\usage{
plco.request(
  resource,
  query = list(),
  api_root = getOption("gwas.explorer.plco.api_root", default =
    "https://exploregwas.cancer.gov/plco-atlas/api"),
  api_key = getOption("gwas.explorer.plco.api_key", default = NULL),
  is_verbose = getOption("gwas.explorer.plco.is_verbose", default = F)
)
}
\arguments{
\item{resource}{A resource name}

\item{query}{A list of query parameters}

\item{api_root}{The API root of the request}

\item{api_key}{An API key to apply to the request}

\item{is_verbose}{If TRUE, enables verbose mode}
}
\value{
The results of the qpi request.
}
\description{
Performs a request against the PLCO API
}
